doble

+++ pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 int _atoi(char *s)
 {
 	int i = 0, sig = 1;
-	long int num = 0;
+	double num = 0;
 
 	while (s[i] != '\0')
 	{
